Small optimizations

+++ pydtnn/initializers/__init__.py
@@ -23,7 +23,7 @@ PyDTNN initializers
 
 import numpy as np
 import scipy.stats as stats
-from enum import Enum, StrEnum, auto
+from enum import StrEnum, auto
 from typing import Callable
 
 class DistributionModeEnum(StrEnum):
@@ -70,7 +70,7 @@ def _generate_distribution(shape: tuple[int, ...], scale:float, mode:Distributio
             x = stats.truncnorm(-2 * stddev, 2 * stddev, loc=0, scale=stddev).rvs(shape).astype(dtype)
         case ProbabilisticDistributionEnum.UNIFORM:
             limit = np.sqrt(3. * scale)
-            x = np.random.uniform(-limit, limit, shape).astype(dtype)
+            x = np.random.uniform(-limit, limit, shape).astype(dtype, copy=True)
         case _:
             raise NotImplementedError(f"distribution: \'{distribution}\' not implemented")
     return x
